test: mock manifest before environment is set up
